fix logic issue

+++ src/Controllers/EmojiController.php
@@ -120,8 +120,8 @@ class EmojiController
      */
     public function updateEmojiByPatch($request, $response, $args)
     {
-        $updateParams = $request->getParsedBody();
-
+        $emoji            = Emoji::find($args['id']);
+        $updateParams     = $request->getParsedBody();
         $validateUserData = $this->authController->validateUserData(['name'], $updateParams);
        
         if (is_array($validateUserData)) {
@@ -132,6 +132,10 @@ class EmojiController
 
         if (is_array($validateArgs)) {
             return $response->withJson($validateArgs, 401);
+        }
+
+        if (is_null($this->getTheOwner($request, $response, $args)->first())) {
+            return ['message' => 'Action cannot be performed because you are not the creator'];
         }
 
         Emoji::where('id', '=', $args['id'])
@@ -151,8 +155,8 @@ class EmojiController
      */
     public function updateEmojiByPut($request, $response, $args)
     {
-        $updateParams = $request->getParsedBody();
-
+        $emoji            = Emoji::find($args['id']);
+        $updateParams     = $request->getParsedBody();
         $validateUserData = $this->authController->validateUserData(['name', 'chars', 'category'], $updateParams);
        
         if (is_array($validateUserData)) {
@@ -163,6 +167,10 @@ class EmojiController
 
         if (is_array($validateArgs)) {
             return $response->withJson($validateArgs, 401);
+        }
+
+        if (is_null($this->getTheOwner($request, $response, $args)->first())) {
+            return ['message' => 'Action cannot be performed because you are not the creator'];
         }
 
         Emoji::where('id', '=', $args['id'])
@@ -182,10 +190,15 @@ class EmojiController
      */
     public function deleteEmoji($request, $response, $args)
     {
+        $emoji        = Emoji::find($args['id']);
         $validateArgs = $this->validateArgs($request, $response, $args);
 
         if (is_array($validateArgs)) {
             return $response->withJson($validateArgs, 401);
+        }
+
+        if (is_null($this->getTheOwner($request, $response, $args)->first())) {
+            return ['message' => 'Action cannot be performed because you are not the creator'];
         }
 
         Emoji::where('id', '=', $args['id'])->delete();
@@ -321,10 +334,6 @@ class EmojiController
         
         if (count($emoji) < 1) {
             return ['message' => 'Action cannot be performed because the id supplied is invalid'];
-        }
-
-        if (is_null($this->getTheOwner($request, $response, $args)->first())) {
-            return ['message' => 'Action cannot be performed because you are not the creator'];
         }
 
         return true;
